+++ packages/address-book-controller/src/AddressBookController.ts
@@ -14,16 +14,14 @@ import {
 import type { Hex } from '@metamask/utils';
 
 /**
- * @type ContactEntry
- *
  * ContactEntry representation
- * @property address - Hex address of a recipient account
- * @property name - Nickname associated with this address
- * @property importTime - Data time when an account as created/imported
  */
 export type ContactEntry = {
+  /** Hex address of a recipient account */
   address: string;
+  /** Nickname associated with this address */
   name: string;
+  /** Data time when an account as created/imported */
   importTime?: number;
 };
 
@@ -31,44 +29,36 @@ export type ContactEntry = {
  * The type of address.
  */
 export enum AddressType {
-  // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/naming-convention
   externallyOwnedAccounts = 'EXTERNALLY_OWNED_ACCOUNTS',
-  // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/naming-convention
   contractAccounts = 'CONTRACT_ACCOUNTS',
-  // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/naming-convention
   nonAccounts = 'NON_ACCOUNTS',
 }
 
 /**
- * @type AddressBookEntry
- *
- * AddressBookEntry representation
- * @property address - Hex address of a recipient account
- * @property name - Nickname associated with this address
- * @property chainId - Chain id identifies the current chain
- * @property memo - User's note about address
- * @property isEns - is the entry an ENS name
- * @property addressType - is the type of this address
+ * AddressBookEntry represents a contact in the address book.
  */
 export type AddressBookEntry = {
+  /** Hex address of a recipient account */
   address: string;
+  /** Nickname associated with this address */
   name: string;
+  /** Chain id identifies the current chain */
   chainId: Hex;
+  /** User's note about address */
   memo: string;
+  /** Indicates if the entry is an ENS name */
   isEns: boolean;
+  /** The type of this address */
   addressType?: AddressType;
+  /** Timestamp of when this entry was last updated */
+  lastUpdatedAt?: number;
 };
 
 /**
- * @type AddressBookState
- *
- * Address book controller state
- * @property addressBook - Array of contact entry objects
+ * State for the AddressBookController.
  */
 export type AddressBookControllerState = {
+  /** Map of chainId to address to contact entries */
   addressBook: { [chainId: Hex]: { [address: string]: AddressBookEntry } };
 };
 
@@ -76,6 +66,12 @@ export type AddressBookControllerState = {
  * The name of the {@link AddressBookController}.
  */
 export const controllerName = 'AddressBookController';
+
+/**
+ * Special chainId used for wallet's own accounts (internal MetaMask accounts).
+ * These entries don't trigger sync events as they are not user-created contacts.
+ */
+const WALLET_ACCOUNTS_CHAIN_ID = '*';
 
 /**
  * The action that can be performed to get the state of the {@link AddressBookController}.
@@ -86,9 +82,53 @@ export type AddressBookControllerGetStateAction = ControllerGetStateAction<
 >;
 
 /**
+ * The action that can be performed to list contacts from the {@link AddressBookController}.
+ */
+export type AddressBookControllerListAction = {
+  type: `${typeof controllerName}:list`;
+  handler: AddressBookController['list'];
+};
+
+/**
+ * The action that can be performed to set a contact in the {@link AddressBookController}.
+ */
+export type AddressBookControllerSetAction = {
+  type: `${typeof controllerName}:set`;
+  handler: AddressBookController['set'];
+};
+
+/**
+ * The action that can be performed to delete a contact from the {@link AddressBookController}.
+ */
+export type AddressBookControllerDeleteAction = {
+  type: `${typeof controllerName}:delete`;
+  handler: AddressBookController['delete'];
+};
+
+/**
+ * Event emitted when a contact is added or updated
+ */
+export type AddressBookControllerContactUpdatedEvent = {
+  type: `${typeof controllerName}:contactUpdated`;
+  payload: [AddressBookEntry];
+};
+
+/**
+ * Event emitted when a contact is deleted
+ */
+export type AddressBookControllerContactDeletedEvent = {
+  type: `${typeof controllerName}:contactDeleted`;
+  payload: [AddressBookEntry];
+};
+
+/**
  * The actions that can be performed using the {@link AddressBookController}.
  */
-export type AddressBookControllerActions = AddressBookControllerGetStateAction;
+export type AddressBookControllerActions =
+  | AddressBookControllerGetStateAction
+  | AddressBookControllerListAction
+  | AddressBookControllerSetAction
+  | AddressBookControllerDeleteAction;
 
 /**
  * The event that {@link AddressBookController} can emit.
@@ -101,7 +141,10 @@ export type AddressBookControllerStateChangeEvent = ControllerStateChangeEvent<
 /**
  * The events that {@link AddressBookController} can emit.
  */
-export type AddressBookControllerEvents = AddressBookControllerStateChangeEvent;
+export type AddressBookControllerEvents =
+  | AddressBookControllerStateChangeEvent
+  | AddressBookControllerContactUpdatedEvent
+  | AddressBookControllerContactDeletedEvent;
 
 const addressBookControllerMetadata = {
   addressBook: { persist: true, anonymous: false },
@@ -159,6 +202,27 @@ export class AddressBookController extends BaseController<
       name: controllerName,
       state: mergedState,
     });
+
+    this.#registerMessageHandlers();
+  }
+
+  /**
+   * Returns all address book entries as an array.
+   *
+   * @returns Array of all address book entries.
+   */
+  list(): AddressBookEntry[] {
+    const { addressBook } = this.state;
+
+    return Object.keys(addressBook).reduce<AddressBookEntry[]>(
+      (acc, chainId) => {
+        const chainIdHex = chainId as Hex;
+        const chainContacts = Object.values(addressBook[chainIdHex]);
+
+        return [...acc, ...chainContacts];
+      },
+      [],
+    );
   }
 
   /**
@@ -188,12 +252,29 @@ export class AddressBookController extends BaseController<
       return false;
     }
 
+    const deletedEntry = { ...this.state.addressBook[chainId][address] };
+
     this.update((state) => {
-      delete state.addressBook[chainId][address];
-      if (Object.keys(state.addressBook[chainId]).length === 0) {
-        delete state.addressBook[chainId];
+      const chainContacts = state.addressBook[chainId];
+      if (chainContacts?.[address]) {
+        delete chainContacts[address];
+
+        // Clean up empty chainId objects
+        if (Object.keys(chainContacts).length === 0) {
+          delete state.addressBook[chainId];
+        }
       }
     });
+
+    // Skip sending delete event for global contacts with chainId '*'
+    // These entries with chainId='*' are the wallet's own accounts (internal MetaMask accounts),
+    // not user-created contacts. They don't need to trigger sync events.
+    if (String(chainId) !== WALLET_ACCOUNTS_CHAIN_ID) {
+      this.messagingSystem.publish(
+        'AddressBookController:contactDeleted',
+        deletedEntry,
+      );
+    }
 
     return true;
   }
@@ -227,8 +308,8 @@ export class AddressBookController extends BaseController<
       memo,
       name,
       addressType,
+      lastUpdatedAt: Date.now(),
     };
-
     const ensName = normalizeEnsName(name);
     if (ensName) {
       entry.name = ensName;
@@ -245,7 +326,35 @@ export class AddressBookController extends BaseController<
       };
     });
 
+    // Skip sending update event for global contacts with chainId '*'
+    // These entries with chainId='*' are the wallet's own accounts (internal MetaMask accounts),
+    // not user-created contacts. They don't need to trigger sync events.
+    if (String(chainId) !== WALLET_ACCOUNTS_CHAIN_ID) {
+      this.messagingSystem.publish(
+        'AddressBookController:contactUpdated',
+        entry,
+      );
+    }
+
     return true;
+  }
+
+  /**
+   * Registers message handlers for the AddressBookController.
+   */
+  #registerMessageHandlers() {
+    this.messagingSystem.registerActionHandler(
+      `${controllerName}:list`,
+      this.list.bind(this),
+    );
+    this.messagingSystem.registerActionHandler(
+      `${controllerName}:set`,
+      this.set.bind(this),
+    );
+    this.messagingSystem.registerActionHandler(
+      `${controllerName}:delete`,
+      this.delete.bind(this),
+    );
   }
 }