+++ packages/network-controller/src/rpc-service/rpc-service-chain.test.ts
@@ -193,22 +193,46 @@ describe('RpcServiceChain', () => {
       };
       // Retry the first endpoint until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint again, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint for a third time, until max retries is hit.
       // The circuit will break on the last time, and the second endpoint will
       // be retried, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
@@ -315,22 +339,54 @@ describe('RpcServiceChain', () => {
       // Retry the first endpoint until max retries is hit.
       await expect(
         rpcServiceChain.request(jsonRpcRequest, fetchOptions),
-      ).rejects.toThrow('Gateway timeout');
+      ).rejects.toThrow(
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
+      );
       // Retry the first endpoint again, until max retries is hit.
       await expect(
         rpcServiceChain.request(jsonRpcRequest, fetchOptions),
-      ).rejects.toThrow('Gateway timeout');
+      ).rejects.toThrow(
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
+      );
       // Retry the first endpoint for a third time, until max retries is hit.
       // The circuit will break on the last time, and the second endpoint will
       // be retried, until max retries is hit.
       await expect(
         rpcServiceChain.request(jsonRpcRequest, fetchOptions),
-      ).rejects.toThrow('Gateway timeout');
+      ).rejects.toThrow(
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
+      );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
       await expect(
         rpcServiceChain.request(jsonRpcRequest, fetchOptions),
-      ).rejects.toThrow('Gateway timeout');
+      ).rejects.toThrow(
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
+      );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
       // The circuit will break on the last time, and the third endpoint will
@@ -415,22 +471,46 @@ describe('RpcServiceChain', () => {
       };
       // Retry the first endpoint until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint again, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint for a third time, until max retries is hit.
       // The circuit will break on the last time, and the second endpoint will
       // be retried, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
@@ -530,22 +610,46 @@ describe('RpcServiceChain', () => {
       };
       // Retry the first endpoint until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint again, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint for a third time, until max retries is hit.
       // The circuit will break on the last time, and the second endpoint will
       // be retried, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
@@ -645,22 +749,46 @@ describe('RpcServiceChain', () => {
       };
       // Retry the first endpoint until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint again, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Retry the first endpoint for a third time, until max retries is hit.
       // The circuit will break on the last time, and the second endpoint will
       // be retried, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.
       await expect(rpcServiceChain.request(jsonRpcRequest)).rejects.toThrow(
-        'Gateway timeout',
+        expect.objectContaining({
+          code: -32002,
+          message: 'RPC endpoint not found or unavailable.',
+          data: {
+            httpStatus: 503,
+          },
+        }),
       );
       // Try the first endpoint, see that the circuit is broken, and retry the
       // second endpoint, until max retries is hit.